-Ultimas alterações feitas em sala, para igualar funções com o cliente -Id do cliente gerado com tempo de maquina -Monitoramento de ação funcionando a cada 5 seg

+++ src/bolsav/StockCli.java
@@ -17,10 +17,10 @@ public class StockCli implements Serializable{
     
     public Stock stock;
     public InterfaceCli client;
-    public int id;
+    public long id;
     public List<InterfaceCli> subscribers;
 
-    public StockCli(Stock stock, InterfaceCli client, int id) {
+    public StockCli(Stock stock, InterfaceCli client, long id) {
         this.stock = stock;
         this.client = client;
         this.id = id;
@@ -35,7 +35,7 @@ public class StockCli implements Serializable{
         return client;
     }
     
-    public int getId() {
+    public long getId() {
         return id;
     }
